+++ dimas/src/agent.rs
@@ -24,20 +24,18 @@ use tokio::time::sleep;
 use zenoh::liveliness::LivelinessToken;
 // endregion:	--- modules
 
-// region:		--- types
-//type AgentProps<P> = std::fmt::Debug + Send + Sync + Unpin + 'static;
-// endregion:	--- types
-
 // region:		--- Agent
 /// implementation of an agent
 pub struct Agent<'a, P>
 where
 	P: std::fmt::Debug + Send + Sync + Unpin + 'static,
 {
+	com: Arc<Communicator>,
 	pd: PhantomData<&'a P>,
+	// The agents propertie structure
+	props: Arc<RwLock<P>>,
 	#[cfg(feature = "liveliness")]
 	liveliness: bool,
-	com: Arc<Communicator>,
 	// an optional liveliness token
 	#[cfg(feature = "liveliness")]
 	liveliness_token: RwLock<Option<LivelinessToken<'a>>>,
@@ -59,8 +57,6 @@ where
 	// registered timer
 	#[cfg(feature = "timer")]
 	timers: Arc<RwLock<Vec<Timer<P>>>>,
-	// The agents propertie structure
-	props: Arc<RwLock<P>>,
 }
 
 impl<'a, P> Agent<'a, P>
@@ -99,12 +95,14 @@ where
 		self.com.uuid()
 	}
 
+	//#[cfg_attr(doc, doc(cfg(feature = "liveliness")))]
 	/// activate sending liveliness information
 	#[cfg(feature = "liveliness")]
 	pub fn liveliness(&mut self, activate: bool) {
 		self.liveliness = activate;
 	}
 
+	//#[cfg_attr(doc, doc(cfg(feature = "liveliness")))]
 	/// get a builder for a subscriber for the liveliness information
 	#[cfg(feature = "liveliness")]
 	pub fn liveliness_subscriber(&self) -> LivelinessSubscriberBuilder<P> {
@@ -113,10 +111,12 @@ where
 			communicator: self.com.clone(),
 			props: self.props.clone(),
 			key_expr: None,
-			callback: None,
+			put_callback: None,
+			delete_callback: None,
 		}
 	}
 
+	//#[cfg_attr(doc, doc(cfg(feature = "subscriber")))]
 	/// get a builder for a Subscriber
 	#[cfg(feature = "subscriber")]
 	pub fn subscriber(&self) -> SubscriberBuilder<P> {
@@ -130,6 +130,7 @@ where
 		}
 	}
 
+	//#[cfg_attr(doc, doc(cfg(feature = "queryable")))]
 	/// get a builder for a Queryable
 	#[cfg(feature = "queryable")]
 	pub fn queryable(&self) -> QueryableBuilder<P> {
@@ -142,6 +143,7 @@ where
 		}
 	}
 
+	//#[cfg_attr(doc, doc(cfg(feature = "publisher")))]
 	/// get a builder for a Publisher
 	#[cfg(feature = "publisher")]
 	pub fn publisher(&self) -> PublisherBuilder<'a, P> {
@@ -153,6 +155,7 @@ where
 		}
 	}
 
+	//#[cfg_attr(doc, doc(cfg(feature = "query")))]
 	/// get a builder for a Query
 	#[cfg(feature = "query")]
 	pub fn query(&self) -> QueryBuilder<P> {
@@ -165,12 +168,10 @@ where
 		}
 	}
 
+	//#[cfg_attr(doc, doc(cfg(feature = "timer")))]
 	/// get a builder for a Timer
 	#[cfg(feature = "timer")]
-	pub fn timer(&self) -> TimerBuilder<P>
-	where
-		P: Default,
-	{
+	pub fn timer(&self) -> TimerBuilder<P> {
 		TimerBuilder {
 			collection: self.timers.clone(),
 			communicator: self.com.clone(),
@@ -185,9 +186,6 @@ where
 	/// # Panics
 	///
 	pub async fn start(&mut self) {
-		// avoid warning
-		let _props = self.props.clone();
-
 		// start all registered queryables
 		#[cfg(feature = "queryable")]
 		self.queryables