+++ dimas/src/com/communicator.rs
@@ -1,20 +1,14 @@
 // Copyright © 2023 Stephan Kunz
 
 // region:		--- modules
-use std::sync::Arc;
-use zenoh::prelude::{r#async::*, sync::SyncResolve};
-
-#[cfg(feature = "query")]
 use super::query::QueryCallback;
-#[cfg(feature = "query")]
 use crate::context::Context;
-#[cfg(feature = "publisher")]
-use crate::prelude::*;
-#[cfg(feature = "query")]
+use crate::error::Result;
+use std::sync::Arc;
 use std::sync::RwLock;
 #[cfg(feature = "liveliness")]
 use zenoh::liveliness::LivelinessToken;
-#[cfg(feature = "publisher")]
+use zenoh::prelude::{r#async::*, sync::SyncResolve};
 use zenoh::publication::Publisher;
 // endregion:	--- modules
 
@@ -47,6 +41,7 @@ impl Communicator {
 		self.prefix.clone()
 	}
 
+	//#[cfg_attr(doc, doc(cfg(feature = "liveliness")))]
 	#[cfg(feature = "liveliness")]
 	pub(crate) async fn liveliness<'a>(
 		&self,
@@ -63,7 +58,6 @@ impl Communicator {
 			.expect("should never happen")
 	}
 
-	#[cfg(feature = "publisher")]
 	pub(crate) async fn create_publisher<'a>(
 		&self,
 		key_expr: impl Into<String> + Send,
@@ -75,13 +69,11 @@ impl Communicator {
 			.expect("should never happen")
 	}
 
-	#[cfg(feature = "publisher")]
 	pub(crate) fn publish<T>(&self, msg_name: impl Into<String>, message: T) -> Result<()>
 	where
-		T: bincode::Encode,
+		T: bitcode::Encode,
 	{
-		let value = bincode::encode_to_vec(message, bincode::config::standard())
-			.expect("should never happen");
+		let value: Vec<u8> = bitcode::encode(&message).expect("should never happen");
 		let key_expr = self.prefix.clone() + "/" + &msg_name.into();
 		//dbg!(&key_expr);
 		match self.session.put(&key_expr, value).res_sync() {
@@ -90,7 +82,6 @@ impl Communicator {
 		}
 	}
 
-	#[cfg(feature = "publisher")]
 	pub(crate) fn delete(&self, msg_name: impl Into<String>) -> Result<()> {
 		let key_expr = self.prefix.clone() + "/" + &msg_name.into();
 		//dbg!(&key_expr);
@@ -100,8 +91,7 @@ impl Communicator {
 		}
 	}
 
-	#[cfg(feature = "query")]
-	pub fn query<P>(
+	pub(crate) fn query<P>(
 		&self,
 		ctx: Arc<Context>,
 		props: Arc<RwLock<P>>,