+++ dimas/src/com/mod.rs
@@ -7,16 +7,12 @@ pub mod communicator;
 #[cfg(feature = "liveliness")]
 pub mod liveliness_subscriber;
 /// Publisher
-#[cfg(feature = "publisher")]
 pub mod publisher;
 /// Query
-#[cfg(feature = "query")]
 pub mod query;
 /// Queryable
-#[cfg(feature = "queryable")]
 pub mod queryable;
 /// Subscriber
-#[cfg(feature = "subscriber")]
 pub mod subscriber;
 // endregion: --- modules