+++ dimas/src/com/subscriber.rs
@@ -9,10 +9,10 @@ use zenoh::prelude::{r#async::AsyncResolve, SampleKind};
 // endregion:	--- modules
 
 // region:		--- types
-/// type definition for a subscribers `publish` callback function
+/// Type definition for a subscribers `publish` callback function
 #[allow(clippy::module_name_repetitions)]
 pub type SubscriberCallback<P> = fn(&Arc<Context>, &Arc<RwLock<P>>, messsage: &[u8]);
-/// type definition for a subscribers `delete` callback function
+/// Type definition for a subscribers `delete` callback function
 pub type DeleteCallback<P> = fn(&Arc<Context>, &Arc<RwLock<P>>);
 // endregion:	--- types
 
@@ -66,12 +66,12 @@ where
 		self
 	}
 
-	/// add the subscriber to the agent
+	/// Build the subscriber
 	/// # Errors
 	///
 	/// # Panics
 	///
-	pub fn add(mut self) -> Result<()> {
+	pub fn build(mut self) -> Result<Subscriber<P>> {
 		if self.key_expr.is_none() {
 			return Err("No key expression or msg type given".into());
 		}
@@ -98,7 +98,19 @@ where
 			props: self.props,
 		};
 
-		self.collection
+		Ok(s)
+	}
+
+	/// Build and add the subscriber to the agent
+	/// # Errors
+	///
+	/// # Panics
+	///
+	pub fn add(self) -> Result<()> {
+		let collection = self.collection.clone();
+		let s = self.build()?;
+
+		collection
 			.write()
 			.expect("should never happen")
 			.push(s);