+++ dimas/src/prelude.rs
@@ -2,11 +2,14 @@
 
 // region:    --- modules
 pub use crate::agent::Agent;
+pub use crate::com::publisher::PublisherBuilder;
+pub use crate::com::query::QueryBuilder;
+pub use crate::com::queryable::{QueryableBuilder, Request};
+pub use crate::com::subscriber::SubscriberBuilder;
 pub use crate::config::Config;
 pub use crate::context::Context;
 pub use crate::error::{Error, Result};
-#[cfg(feature = "queryable")]
-pub use crate::com::queryable::Request;
+pub use crate::timer::TimerBuilder;
 // endregeion:  --- modules
 
 // region:    --- types